requirements.txt
ecdsa==0.19.0
pyasn1==0.6.1
pycryptodome==3.21.0
rsa==4.9
six==1.17.0
